created recipe page for mac and cheese and anchor
Create mac and cheese recipe page to detail the mac and cheese recipe
with description of recipe, the ingredients, and steps to cooke it.
There is also an image in the image directory that shows the final
look of the mac and cheese. I also updated the anchor in index.html
to link to the new recipe.